requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.4.1
sentence-transformers==2.5.1
groq==0.4.2
python-dotenv==1.0.1
pydantic==2.6.4
